Playing with different styles. Lot of new junk code.
