+++ package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "chalk": "^2.3.0",
+    "is-empty": "^1.2.0",
     "multiline": "^1.0.2"
   },
   "devDependencies": {
