+++ src/content.js
@@ -3,62 +3,43 @@
  * @Date:   2018-01-24T16:32:34-08:00
  * @Email:  alec@bubblegum.academy
  * @Last modified by:   alechp
- * @Last modified time: 2018-01-25T13:04:32-08:00
+ * @Last modified time: 2018-01-25T16:43:11-08:00
  */
 const log = console.log;
 const path = require("path");
 const chalk = require("chalk");
+const empty = require("is-empty");
 const { fluent } = require("./helpers.js");
 
-function Content(template, variables) {
-  this.template = template;
-  this.variables = variables;
-}
+/**
+ * Interpolation for templating
+ * @param {[string]} template  [multiline template string generated by github.com/sindresorhus/multiline]
+ * @param {[object]} variables [variables being interpolated, eg. {foo: 'bar'}]
+ */
+function interpolate(template, variables) {}
 
-Content.prototype.interpolate = fluent(function() {
-  let temp = this.templateStringGenerator(this.template);
-  log(temp(this.variables));
-});
-
-Content.prototype.templateStringGenerator = fluent(function() {
-  /*
-
-
-  var kingMaker = generateTemplateString('${name} is king!');
-
-  console.log(kingMaker({name: 'Bryan'}));
-  // Logs 'Bryan is king!' to the console.
-
-  */
-  var cache = {};
-
-  function generateTemplate() {
-    var fn = cache[this.template];
-    log(`This inside generateTemplate: ${chalk.yellow(this)}`);
-    if (!fn) {
-      // Replace ${expressions} (etc) with ${map.expressions}.
-
-      var sanitized = this.template
-        .replace(/\$\{([\s]*[^;\s\{]+[\s]*)\}/g, function(_, match) {
-          return `\$\{map.${match.trim()}\}`;
-        })
-        // Afterwards, replace anything that's not ${map.expressions}' (etc) with a blank string.
-        .replace(/(\$\{(?!map\.)[^}]+\})/g, "");
-
-      log(`This inside generateTemplate if block: ${chalk.yellow(this)}`);
-      fn = Function("map", `return \`${sanitized}\``);
-    }
-
-    return fn;
-  }
-
-  return generateTemplate; // returns reference, doesn't invoke
-});
-
-function interpolate(template) {
-  const c = new Content(template);
-  c.templateStringGenerator().interpolate();
-}
-module.exports = {
-  interpolate
-};
+// let template = () => {
+//   let t = this.template;
+//   return (t) => {
+//   let sanitized = t
+//     .replace(/\$\{([\s]*[^;\s\{]+[\s]*)\}/g, function(_, match) {
+//       return `\$\{map.${match.trim()}\}`;
+//     })
+//     // Afterwards, replace anything that's not ${map.expressions}' (etc) with a blank string.
+//     .replace(/(\$\{(?!map\.)[^}]+\})/g, "");
+//   return sanitized;
+// }
+// return Function("map", `return \`${sanitized}\``); // whats with backticks yo ?
+// let templateGenerator = template => {};
+// const play = () => {
+//   let _t = this;
+//
+//   return {
+//     thising: () => {
+//       log(_t);
+//       log(this);
+//     }
+//   };
+// };
+//
+// play().thising();
